requirements.txt
fastapi
uvicorn
pymongo
transformers
torch
scikit-learn
joblib
python-dotenv
